event-driven-system/user_service_v1/requirements.txt
fastapi
uvicorn
pydantic
pymongo
pika
python-dotenv
pydantic[email]
email-validator
